requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
anyio==3.7.1
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.2
chromadb==0.4.5
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
dashscope==1.14.1
dataclasses-json==0.6.4
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.99.1
fastjsonschema==2.19.1
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.1
humanfriendly==10.0
idna==3.6
importlib_resources==6.1.2
ipython==8.12.3
jedi==0.19.1
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
langchain==0.0.341
langchain-core==0.0.6
langsmith==0.0.87
lxml==5.1.0
MarkupSafe==2.1.5
marshmallow==3.21.0
matplotlib-inline==0.1.6
mistune==3.0.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
openai==1.13.3
openpyxl==3.1.2
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pipdeptree==2.15.1
pipreqs==0.5.0
platformdirs==4.2.0
posthog==3.4.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
pydantic==1.10.12
Pygments==2.17.2
pypdf==3.16.2
PyPika==0.48.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rpds-py==0.18.0
safetensors==0.4.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.27
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.1
tinycss2==1.2.1
tokenizers==0.15.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
transformers==4.38.1
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.27.1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==12.0
yarg==0.1.9
yarl==1.9.4
